# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a covidus server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
        "version": "1.0.0",
        "title": "SCovidus",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "covidus1.herokuapp.com",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
       "Bearer":{
           "type":"apiKey",
           "description": " pls login",
           "in": "header",
           "name": "Authorization"
       }
    },
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "Root route"
                ],
                "summary": "get to home route",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definition/home"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Authenticated route for user"
                ],
                "description": "This is an anthenticated route after user user signs in",
                "summary": "User home",
                "produces": [
                    "application/json"
                ],
                "security": [
                    {
                        "Bearer" : []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401":{
                        "description": "Empty token"
                    },
                    "400":{
                        "description": "Expire token"
                    }
                }
            }
        },
        "/covid-info": {
            "get": {
                "tags": [
                    "Covid-19 info For Countries"
                ],
                "summary": "Covid-19 restrictions for countries",
                "description": "to get countries and their their covid restrictions",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Covid-info"
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "number"
                        },
                        "description": "Query the number of data"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "type": "string",
                        "description": "A search query"
                    }
                ]
            }
        },
        "/searchCovidInfo": {
            "get": {
                "tags": [
                    "Covid search"
                ],
                "description": "Search for information that relates to covid-19 restriction in country and the search is based on country name which is /adm0_name/ any thing other than that will return empty",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Covid-infoh"
                        }
                    },
                    "404": {
                        "description": "failed operation"
                    }
                },
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "type": "string",
                        "description": "A search query"
                    }
                ]
            }
        },
        "/more-video": {
            "get": {
                "tags": [
                    "Get more videos"
                ],
                "description": "get more videos on click buuton",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": " successfull response"
                    }
                }
            }
        },
        "/postVideo": {
            "post": {
                "tags": [
                    "videos upload"
                ],
                "description": "upload media with  extension of .mp4 .mp3 etc",
                "operationId": "uploadFile",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "multipart/json"
                ],
                "security": [
                    {
                        "Bearer":[]
                    }
                ],
                "parameters": [
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "file to upload",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "content",
                        "in": "formData",
                        "description": "Additional data to pass to server",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name":"county",
                        "in":"formData",
                        "description": "Additional file to pass to server",
                        "required": false,
                        "type":"string"
                    }
                    

                ],
                "responses": {
                    "200": {
                        "description": "image upload successfully",
                        "schema": {
                            "$ref": "#/definitions/Videoupload"
                        }
                    },
                    "404":{
                        "description": "authentication error"
                    },
                    "405":{
                        "description": "invalid video format"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "description": "login",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",

                        "schema": {
                            "$ref": "#/definitions/Login"
                        }
                    },
                    "400":{
                        "description": " empty request body"
                    },
                    "401":{
                        "description": "Invalid password or email"
                    }
                }
            }
        },
        "/auth/facebook": {
            "get": {
                "tags": [
                    "Facebook authentication"
                ],
                "summary": "Login or signup link  with facebook",
                "description": "After successfull signin or signup the user will be redirect to /user route",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "image upload successfully"
                    }
                }
            }
        },
        "/auth/google": {
            "get": {
                "tags": [
                    "Facebook authentication"
                ],
                "summary": "Authentication with Google",
                "description": "After a successfull google authentication the user will be redirected to /user",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "image upload successfully",
                        "schema": {
                            "$ref": "#/definitions/Login"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "tags": [
                    "Signup"
                ],
                "description": "signup with email , name",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Signup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfull operation",
                        "schema": {
                            "$ref": "#/definitions/Signup"
                        }
                    },
                    "400":{
                        "description": "errro in validation"
                    },
                    "401": {
                        "description": "server error"
                      },
                      "404": {
                        "description": "Validation exception"
                      }
                }
            }
        },
        "/contact": {
            "get": {
                "tags": [
                    "Contact us page"
                ],
                "summary": "contact us page",
                "description": "get route for contact us page",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successfull operation"
                    }
                }
            }
        },
        "/contact/sendMessage": {
            "post": {
                "tags": [
                    "Send Messsage"
                ],
                "summary": "Post route for contact us ",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ContactSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfull operation",
                        "schema": {
                            "$ref": "#/definitions/ContactSchema"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "home": {
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
        "Covid-info": {
            "type": "object",
            "properties": {
                "adm0_name": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "iso3": {
                    "type": "string"
                },
                "ObjectId": {
                    "type": "string"
                },
                "optional1": {
                    "type": "string"
                },
                "optional2": {
                    "type": "string"
                },
                "optional3": {
                    "type": "string"
                },
                "sources": {
                    "type": "string"
                }
            }
        },
        "Videoupload": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                }
            }
        },
        "Login": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "Signup": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "ContactSchema": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "number"
                },
                "reason": {
                    "type": "string"
                },
                "evidence": {
                    "type": "string"
                }
            }
        }
    }
}